requirements.txt
flask
gunicorn
sqlalchemy
pymysql
python-dotenv